Update gradient computations. Implement rough gradient check. Add some comments and an error check. Use double instead of float. Remove plotting functionality.

+++ Data.h
@@ -3,6 +3,6 @@
 class Data
 {
 	public:
-		Eigen::MatrixXf examples;
+		Eigen::MatrixXd examples;
 		Eigen::VectorXi labels;
 };
